Replace "redisKey" method with the "RedisKeys" class

+++ discord/loritta-bot-common/src/main/kotlin/net/perfectdreams/loritta/morenitta/utils/newconfig/LorittaConfig.kt
@@ -104,8 +104,8 @@ data class LorittaConfig(
         class LorittaClusterConfig(
             val id: Int,
             val name: String,
-            val minShard: Long,
-            val maxShard: Long,
+            val minShard: Int,
+            val maxShard: Int,
             val websiteUrl: String
         ) {
             fun getUrl(loritta: LorittaBot) = DiscordUtils.getUrlForLorittaClusterId(loritta, id)
